* background.c (background_attention): Cast pointer to long instead of int because it's unsafe of 64-bit systems. * main.c (init_xterm_support): Likewise.

+++ src/background.c
@@ -285,7 +285,7 @@ background_attention (int fd, void *closure)
     }
 
     /* If the routine is zero, then the child is telling us that he is dying */
-    if ((int) routine == MSG_CHILD_EXITING){
+    if ((long) routine == MSG_CHILD_EXITING){
 	unregister_task_running (ctx->pid, fd);
 	waitpid (ctx->pid, &status, 0);
 	return 0;
